+++ Season 9/service/CalculatorService.java
@@ -4,5 +4,5 @@ public interface CalculatorService{
 	public Integer tambah (int bil1, int bil2);
 	public Integer kurang (int bil1, int bil2);
 	public Integer kali (int bil1, int bil2);
-	public Integer bagi (int bil1, int bil2);
+	public Double bagi (double bil1, double bil2);
 }
